Перерапределил методы по соответствующим контроллерам и сервисам.
